requirements.txt
Flask
Flask-WTF
requests
dnspython
python-whois
ipwhois
python-dotenv
scapy